moved states-machine logic to a different file
